Additional spec preventing deleting of homepage, removed the specs for the right content

+++ wiki.rb
@@ -89,7 +89,7 @@ class SimpleWiki < Sinatra::Base
 
   post '/:page' do |page|
     fname = File.join(settings.views,"#{page}.md")
-    if params[:content].empty?
+    if params[:content].empty? and page != 'homepage'
       File.delete(fname)
       redirect '/'
     else
